+++ apps/lambda_function_scheduled/lambda.tf
@@ -1,0 +1,106 @@
+resource "aws_lambda_function" "app" {
+  filename         = "${var.filename}"
+  function_name = "${var.app_name}"
+  role = "${aws_iam_role.iam_for_app.arn}"
+  handler = "${var.handler}"
+  vpc_config = {
+    subnet_ids = "${var.private == true ? ${module.aws_core_data.private_subnets} : ${concat(module.aws_core_data.private_subnets,module.aws_core_data.public_subnets)}}"
+    security_group_ids = ["${aws_security_group.sg_for_app.id}"]
+  }
+  count            = "${var.enabled}"
+}
+
+resource "aws_lambda_permission" "cloudwatch" {
+  statement_id  = "AllowExecutionFromCloudWatch"
+  action        = "lambda:InvokeFunction"
+  function_name = "${aws_lambda_function.app.arn}"
+  principal     = "events.amazonaws.com"
+  source_arn    = "${aws_cloudwatch_event_rule.app.arn}"
+  count         = "${var.enabled}"
+}
+
+resource "aws_iam_role" "iam_for_app" {
+  name = "${var.app_name}"
+  assume_role_policy = <<EOF
+{
+  "Version": "2012-10-17",
+  "Statement": [
+    {
+      "Action": "sts:AssumeRole",
+      "Principal": {
+        "Service": "lambda.amazonaws.com"
+      },
+      "Effect": "Allow",
+      "Sid": ""
+    }
+  ]
+}
+EOF
+}
+
+resource "aws_iam_role_policy" "iam_policy_for_app" {
+  name = "${var.app_name}"
+  role = "${aws_iam_role.iam_for_app.id}"
+  policy = <<EOF
+{
+  "Version": "2012-10-17",
+  "Statement": [
+        {
+            "Effect": "Allow",
+            "Action": [
+                "logs:CreateLogGroup",
+                "logs:CreateLogStream",
+                "logs:PutLogEvents"
+            ],
+            "Resource": "arn:aws:logs:*:*:*"
+        },
+    {
+      "Effect": "Allow",
+      "Action": [
+        "ec2:CreateNetworkInterface"
+      ],
+      "Resource": [
+        "*"
+      ]
+    }
+  ]
+}
+EOF
+}
+
+resource "aws_security_group" "sg_for_app" {
+  name = "${var.app_name}"
+  description = "Allow all inbound traffic for the scheduled lambda function"
+  vpc_id = "${module.aws_core_data.vpc_id}"
+
+  ingress {
+    from_port = 0
+    to_port = 0
+    protocol = "-1"
+    cidr_blocks = ["0.0.0.0/0"]
+  }
+
+  egress {
+    from_port = 0
+    to_port = 0
+    protocol = "-1"
+    cidr_blocks = ["0.0.0.0/0"]
+  }
+
+  tags {
+    Name            = "${var.app_name}"
+    "Business Unit" = "${var.tags_business_unit}"
+    "Cost Center"   = "${var.tags_cost_center}"
+    Team            = "${var.tags_team}"
+    Purpose         = "${var.tags_purpose}"
+    Description     = "${var.description}"
+    Environment     = "${var.environment}"
+  }
+}
+
+//resource "aws_lambda_alias" "app_alias" {
+//  name = "${var.app_name}"
+//  description = "${var.description}"
+//  function_name = "${aws_lambda_function.app.arn}"
+//  function_version = "$LATEST"
+//}