+++ apps/lambda_function_scheduled/main.tf
@@ -1,0 +1,7 @@
+provider "aws" {
+  region = "${var.region}"
+}
+
+module "aws_core_data" {
+  source = "git@github.com:albumprinter/eops_tf_modules.git//data_providers/aws_core_data?ref=v0.1.1"
+}